refactor: [FE] Editor 컴포넌트 EditorSection으로 이동 기존 Editor의 역할을 EditorSection이 수행하도록 변경
