+++ frontEnd/src/components/room/editor/EditorHeader.tsx
@@ -1,0 +1,7 @@
+interface EditorFileNameProps {
+  children: React.ReactNode;
+}
+
+export default function EditorFileName({ children }: EditorFileNameProps) {
+  return <h1 className="text-white text-[max(2vh,15px)]">{children}</h1>;
+}
